refactoring: move deviceName, timeFormat, formatAgo, timeAt to utils
