+++ lib/client/browser-settings.js
@@ -13,7 +13,7 @@ function init (client, serverSettings, $) {
   var settings = require('../settings')();
 
   function loadForm ( ) {
-    var utils = require('../utils')(settings);
+    var utils = client.utils;
     var language = require('../language')();
     language.set(settings.language);
     var translate = language.translate;
