+++ lib/client/index.js
@@ -149,7 +149,10 @@ client.load = function load(serverSettings, callback) {
     , pluginBase: client.plugins.base(majorPills, minorPills, statusPills, bgStatus, client.tooltip, $.localStorage)
   };
 
-  client.utils = require('../utils')(client.ctx.settings);
+  client.utils = require('../utils')({
+    settings: client.settings
+    , language: language
+  });
   client.sbx = sandbox.clientInit(client.ctx, client.now);
   client.renderer = require('./renderer')(client, d3, $);
