+++ lib/plugins/bgnow.js
@@ -10,7 +10,7 @@ var bucketFields = ['index', 'fromMills', 'toMills'];
 function init (ctx) {
 
   var translate = ctx.language.translate;
-  var timeago = require('./timeago')(ctx);
+  var utils = require('../utils')(ctx);
 
   var bgnow = {
     name: 'bgnow'
@@ -196,9 +196,9 @@ function init (ctx) {
 
     if (prop.sgvs && prop.sgvs.length > 1) {
       _.forEach(prop.sgvs, function deviceAndValue(entry) {
-        var device = deviceName(entry.device);
+        var device = utils.deviceName(entry.device);
         deviceInfos[device] = {
-          time: timeFormat(moment(entry.mills), sbx)
+          time: utils.timeFormat(moment(entry.mills), sbx)
           , value: sbx.scaleEntry(entry)
           , recent: entry
         };
@@ -207,7 +207,7 @@ function init (ctx) {
 
     if (delta && delta.previous && delta.previous.sgvs && delta.previous.sgvs.length > 1) {
       _.forEach(delta.previous.sgvs, function deviceAndValue(entry) {
-        var device = deviceName(entry.device);
+        var device = utils.deviceName(entry.device);
         var deviceInfo = deviceInfos[device];
         if (deviceInfo) {
           var deviceDelta = bgnow.calcDelta(
@@ -221,7 +221,7 @@ function init (ctx) {
           }
         } else {
           deviceInfos[device] = {
-            time: timeFormat(moment(entry.mills), sbx)
+            time: utils.timeFormat(moment(entry.mills), sbx)
             , value: sbx.scaleEntry(entry)
           };
         }
@@ -250,29 +250,6 @@ function init (ctx) {
       , info: _.isEmpty(info) ? null : info
     });
   };
-
-  function deviceName (device) {
-    var last = device ? _.last(device.split('://')) : 'unknown';
-    return _.first(last.split('/'));
-  }
-
-  function timeFormat (m, sbx) {
-    var when;
-    if (m && sbx.data.inRetroMode) {
-      when = m.format('LT');
-    } else if (m) {
-      when = formatAgo(m, sbx.time);
-    } else {
-      when = 'unknown';
-    }
-
-    return when;
-  }
-
-  function formatAgo (m, nowMills) {
-    var ago = timeago.calcDisplay({mills: m.valueOf()}, nowMills);
-    return translate('%1' + ago.shortLabel + (ago.shortLabel.length === 1 ? ' ago' : ''), { params: [(ago.value ? ago.value : '')]});
-  }
 
   return bgnow;
