+++ lib/plugins/iob.js
@@ -2,12 +2,12 @@
 
 var _ = require('lodash')
   , moment = require('moment')
-  , times = require('../times')
-  , utils = require('../utils')();
+  , times = require('../times');
 
 function init(ctx) {
   var translate = ctx.language.translate;
-
+  var utils = require('../utils')(ctx);
+  
   var iob = {
     name: 'iob'
     , label: 'Insulin-on-Board'
