+++ lib/plugins/loop.js
@@ -9,6 +9,7 @@ var levels = require('../levels');
 
 function init(ctx) {
   var timeago = require('./timeago')(ctx);
+  var utils = require('../utils')(ctx);
 
   var loop = {
     name: 'loop'
@@ -153,7 +154,7 @@ function init(ctx) {
       sbx.notifications.requestNotify({
         level: level
         , title: 'Loop isn\'t looping'
-        , message: 'Last Loop: ' + formatAgo(prop.lastLoop.moment, now.valueOf())
+        , message: 'Last Loop: ' + utils.formatAgo(prop.lastLoop.moment, now.valueOf())
         , pushoverSound: 'echo'
         , group: 'Loop'
         , plugin: loop
@@ -291,7 +292,7 @@ function init(ctx) {
 
     var info = _.map(sorted, function eventToInfo (event) {
       return {
-        label: timeAt(false, sbx) + timeFormat(event.time, sbx)
+        label: utils.timeAt(false, sbx) + utils.timeFormat(event.time, sbx)
         , value: event.value
       };
     });
@@ -307,7 +308,7 @@ function init(ctx) {
     var lastLoopMoment = prop.lastLoop ? prop.lastLoop.moment : null;
 
     sbx.pluginBase.updatePillText(loop, {
-      value: timeFormat(lastLoopMoment, sbx)
+      value: utils.timeFormat(lastLoopMoment, sbx)
       , label: label
       , info: info
       , pillClass: statusClass(prop, prefs, sbx)
@@ -400,29 +401,6 @@ function init(ctx) {
     }
 
     return level;
-  }
-
-  function timeFormat (m, sbx) {
-
-    var when;
-    if (m && sbx.data.inRetroMode) {
-      when = m.format('LT');
-    } else if (m) {
-      when = formatAgo(m, sbx.time);
-    } else {
-      when = 'unknown';
-    }
-
-    return when;
-  }
-
-  function formatAgo (m, nowMills) {
-    var ago = timeago.calcDisplay({mills: m.valueOf()}, nowMills);
-    return (ago.value ? ago.value : '') + ago.shortLabel + (ago.shortLabel.length === 1 ? ' ago' : '');
-  }
-
-  function timeAt (prefix, sbx) {
-    return sbx.data.inRetroMode ? (prefix ? ' ' : '') + '@ ' : (prefix ? ', ' : '');
   }
 
   return loop;
