+++ lib/plugins/openaps.js
@@ -9,7 +9,7 @@ var levels = require('../levels');
 
 function init(ctx) {
   var timeago = require('./timeago')(ctx);
-  var translate = ctx.language.translate;
+  var utils = require('../utils')(ctx);
 
   var openaps = {
     name: 'openaps'
@@ -96,7 +96,7 @@ function init(ctx) {
 
       if (!device) {
         device = {
-          name: uri.indexOf('openaps://') === 0 ? uri.substring('openaps://'.length) : uri
+          name: utils.deviceName(uri)
           , uri: uri
         };
 
@@ -264,7 +264,7 @@ function init(ctx) {
       sbx.notifications.requestNotify({
         level: level
         , title: 'OpenAPS isn\'t looping'
-        , message: 'Last Loop: ' + formatAgo(prop.lastLoopMoment, now.valueOf())
+        , message: 'Last Loop: ' + utils.formatAgo(prop.lastLoopMoment, now.valueOf())
         , pushoverSound: 'echo'
         , group: 'OpenAPS'
         , plugin: openaps
@@ -420,7 +420,7 @@ function init(ctx) {
 
     var info = _.map(sorted, function eventToInfo (event) {
       return {
-        label: timeAt(false, sbx) + timeFormat(event.time, sbx)
+        label: utils.timeAt(false, sbx) + utils.timeFormat(event.time, sbx)
         , value: event.value
       };
     });
@@ -431,7 +431,7 @@ function init(ctx) {
     }
 
     sbx.pluginBase.updatePillText(openaps, {
-      value: timeFormat(prop.lastLoopMoment, sbx)
+      value: utils.timeFormat(prop.lastLoopMoment, sbx)
       , label: label
       , info: info
       , pillClass: statusClass(prop, prefs, sbx)
@@ -491,29 +491,6 @@ function init(ctx) {
     }
 
     return level;
-  }
-
-  function timeFormat (m, sbx) {
-
-    var when;
-    if (m && sbx.data.inRetroMode) {
-      when = m.format('LT');
-    } else if (m) {
-      when = formatAgo(m, sbx.time);
-    } else {
-      when = 'unknown';
-    }
-
-    return when;
-  }
-
-  function formatAgo (m, nowMills) {
-    var ago = timeago.calcDisplay({mills: m.valueOf()}, nowMills);
-    return translate('%1' + ago.shortLabel + (ago.shortLabel.length === 1 ? ' ago' : ''), { params: [(ago.value ? ago.value : '')]});
-  }
-
-  function timeAt (prefix, sbx) {
-    return sbx.data.inRetroMode ? (prefix ? ' ' : '') + '@ ' : (prefix ? ', ' : '');
   }
 
   return openaps;
