+++ lib/report_plugins/index.js
@@ -38,8 +38,13 @@ function init() {
 
   plugins.consts = consts;
 
-  plugins.utils = require('./utils')();
-  
+  plugins.utils = require('./utils')({
+    //TODO: refactor so all this happens after init somehow
+    //  until then use some defaults so things don't blow up
+    language: require('../language')()
+    , settings: {}
+  });
+
   plugins.addHtmlFromPlugins = function addHtmlFromPlugins(client) {
     plugins.eachPlugin(function addHtml(p) {
       // add main plugin html
