+++ lib/utils.js
@@ -1,10 +1,15 @@
 'use strict';
 
+var _ = require('lodash');
 var moment = require('moment-timezone');
 
 var units = require('./units')();
 
-function init(settings) {
+function init(ctx) {
+
+  var settings = ctx.settings;
+  var translate = ctx.language.translate;
+  var timeago = require('./plugins/timeago')(ctx);
 
   var utils = { };
 
@@ -34,7 +39,35 @@ function init(settings) {
     return moment(datestring + ' ' + timestring, 'YYYY-MM-D HH:mm');
   };
 
- return utils;
+
+  utils.deviceName = function deviceName (device) {
+    var last = device ? _.last(device.split('://')) : 'unknown';
+    return _.first(last.split('/'));
+  };
+
+  utils.timeFormat = function timeFormat (m, sbx) {
+    var when;
+    if (m && sbx.data.inRetroMode) {
+      when = m.format('LT');
+    } else if (m) {
+      when = utils.formatAgo(m, sbx.time);
+    } else {
+      when = 'unknown';
+    }
+
+    return when;
+  };
+
+  utils.formatAgo = function formatAgo (m, nowMills) {
+    var ago = timeago.calcDisplay({mills: m.valueOf()}, nowMills);
+    return translate('%1' + ago.shortLabel + (ago.shortLabel.length === 1 ? ' ago' : ''), { params: [(ago.value ? ago.value : '')]});
+  };
+
+  utils.timeAt = function timeAt (prefix, sbx) {
+    return sbx.data.inRetroMode ? (prefix ? ' ' : '') + '@ ' : (prefix ? ', ' : '');
+  };
+
+  return utils;
 }
 
 module.exports = init;
